+++ lib/dart_pusher_channels.dart
@@ -2,3 +2,5 @@ library dart_pusher_channels;
 
 export 'base.dart';
 export 'api.dart';
+
+const kDartPusherChannelsLibraryVersion = '0.2.8';